+++ CHANGELOG.md
@@ -52,6 +52,7 @@ Expected: September 2022
 
 ### Corrected Bugs
 
+* Fixed: [YANG min-elements within non-presence container does not work](https://github.com/clicon/clixon/issues/355)
 * Fixed: [Issues with ietf-snmp modules](https://github.com/clicon/clixon/issues/353)
 * Fixed: [Missing/no namespace error in YANG augments with default values](https://github.com/clicon/clixon/issues/354)
 * Fixed: [Validation of mandatory in choice/case does not work in some cases](https://github.com/clicon/clixon/issues/349)